refactor(python): move tl + graphrag into infomap.tl package
git mv tl.py -> tl/__init__.py (preserves the infomap.tl.infomap(...) anndata
API) and graphrag.py -> tl/graphrag.py. tl/__init__ now imports Infomap via
'..\_facade'; graphrag keeps its absolute 'from infomap import Infomap'.

A back-compat shim infomap/graphrag.py re-exports the public graphrag API plus
the private _output_paths from infomap.tl.graphrag, so
'from infomap.graphrag import read_graphrag, write_graphrag_communities,
run_graphrag_communities, _output_paths' keeps working (test_graphrag.py).
merge.py and shell.py stay put (python -m infomap.merge / .shell).
